Added dialog for rewards
